Revert signSchnorr to use default all-zero Buffer for auxRand
This commit reverts the behavior of the `signSchnorr` function to use an all-zero Buffer when `auxRand` isn't provided, in alignment with the specifications outlined by `bitcoinjs-lib` for `secp256k1`. This update ensures consistency between the two libraries and prevents discrepancies in Schnorr signatures. Additionally, version number is incremented to 1.0.5 to reflect this change.